Fix JSR/JMP to use CALL Pcode where possible. (#38)
This means JSR/JMP to non-computed destinations are now resolved to references at disassembly time, rather than analysis time. This, in turn, results in the destination function being the primary reference for the instruction, regardless of whether the instruction later acquires a reference to the stack where the return address is written.

Also tidy some of the spec by eliminating explicit temporaries here and there.